+++ Core/SoarKernel/src/rete.cpp
@@ -6444,7 +6444,7 @@ void reteload_all_symbols (agent* thisAgent, FILE* f) {
   for (i=0; i<num_float_constants; i++) {
     reteload_string(f);
     *(current_place_in_symtab++) = 
-      make_float_constant (thisAgent, (float)my_strtod(reteload_string_buf,NULL,10));
+      make_float_constant (thisAgent, (double)my_strtod(reteload_string_buf,NULL,10));
   }
 }
